Separate out custom code into module
Separates custom runners and parameter sources into a separate module so
that they can be shared across different tracks.
