+++ benchmarks/osb/helpers/helper.py
@@ -1,0 +1,32 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# The OpenSearch Contributors require contributions made to
+# this file be licensed under the Apache-2.0 license or a
+# compatible open source license.
+
+import numpy as np
+from typing import List
+from typing import Dict
+from typing import Any
+
+
+def bulk_transform(partition: np.ndarray, field_name: str, action,
+                   offset: int) -> List[Dict[str, Any]]:
+    """Partitions and transforms a list of vectors into OpenSearch's bulk
+    injection format.
+    Args:
+        offset: to start counting from
+        partition: An array of vectors to transform.
+        field_name: field name for action
+        action: Bulk API action.
+    Returns:
+        An array of transformed vectors in bulk format.
+    """
+    actions = []
+    _ = [
+        actions.extend([action(i + offset), None])
+        for i in range(len(partition))
+    ]
+    actions[1::2] = [{field_name: vec} for vec in partition.tolist()]
+    return actions
+
